requirements.txt
astroid==2.0.2
autopep8==1.3.5
colorama==0.3.9
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.4.0
pycryptodome==3.6.4
pylint==2.1.0
six==1.11.0
typed-ast==1.1.0
typing==3.6.4
wrapt==1.10.11